Anpassung an Architektur
Hier wurden nun die bisherigen Klassen umbenannt (auch ihre Dateinamen
selbst) und teilweise auch entfernt, da sie nicht der Architektur
entsprachen.
