--- Würfel_Button.java
@@ -1,19 +1,0 @@
-import greenfoot.*;  // (World, Actor, GreenfootImage, Greenfoot and MouseInfo)
-
-/**
- * Write a description of class Würfel_Button_Angriff here.
- * 
- * @author (your name) 
- * @version (a version number or a date)
- */
-public class Würfel_Button extends Button
-{
-    /**
-     * Act - do whatever the Würfel_Button_Angriff wants to do. This method is called whenever
-     * the 'Act' or 'Run' button gets pressed in the environment.
-     */
-    public void act() 
-    {
-        // Add your action code here.
-    }    
-}
